+++ Interface/AddOns/BadBoy/CHANGELOG.md
@@ -1,7 +1,8 @@
 # BadBoy
 
-## [v7.1.13](https://github.com/funkydude/BadBoy/tree/v7.1.13) (2016-11-04) [](#top)
-[Full Changelog](https://github.com/funkydude/BadBoy/compare/v7.1.12...v7.1.13)
+## [v7.1.14](https://github.com/funkydude/BadBoy/tree/v7.1.14) (2016-11-06) [](#top)
+[Full Changelog](https://github.com/funkydude/BadBoy/compare/v7.1.13...v7.1.14)
 
 - anti-spam update  
+- Tweak button, clarify how to dismiss spam in the tooltip.  
 - anti-spam update