+++ Interface/AddOns/LittleWigs/CHANGELOG.md
@@ -1,7 +1,8 @@
 # LittleWigs
 
-## [v7.1.1-2-g9c7eb42](https://github.com/BigWigsMods/LittleWigs/tree/9c7eb42434860bbd4628d63916a51cc831e1decf) (2016-11-06) [](#top)
-[Full Changelog](https://github.com/BigWigsMods/LittleWigs/compare/v7.1.1...9c7eb42434860bbd4628d63916a51cc831e1decf)
+## [v7.1.1-3-g60d8a10](https://github.com/BigWigsMods/LittleWigs/tree/60d8a1007fa66d467def8bc311c91ee64771a448) (2016-11-07) [](#top)
+[Full Changelog](https://github.com/BigWigsMods/LittleWigs/compare/v7.1.1...60d8a1007fa66d467def8bc311c91ee64771a448)
 
+- Karazhan/MaidenOfVirtue: Fix Repentance id  
 - Add Maiden (#7)  
 - BlackRookHold/Smashspite: Add warning for Hateful Gaze, cleanups.